splits the use of axios for rest and frontend
