+++ src/rest/axios.js
@@ -5,7 +5,6 @@ const JSONBigInt = require('json-bigint')({ 'storeAsString': true })
 
 let instance
 let limiter
-const cookies = new Map()
 
 const handleBigInt = (data) => {
   try {
@@ -30,9 +29,7 @@ function _exec ({ method, apiKey, target = 'api', url = '/', data, params = {} }
     maxRedirects,
     headers: {
       'X-Cisco-Meraki-API-Key': apiKey,
-      maxredirects: maxRedirects,
-      'Cookie': Array.from(cookies).map((cookie) => `${cookie[0]}=${cookie[1]}`).join(';'),
-      accept: 'application/json'
+      maxredirects: maxRedirects
     },
     withCredentials: true,
     params,
@@ -63,15 +60,7 @@ function _delete (apiKey, target, url) {
   return _exec({ method: 'DELETE', apiKey, target, url })
 }
 
-function setCookie (key, value) {
-  cookies.set(key, value)
-}
-
-function deleteCookie (key) {
-  cookies.delete(key)
-}
-
-module.exports = ({ baseUrl, rateLimiter }) => {
+module.exports = ({ baseUrl, rateLimiter, headers }) => {
   if (!instance) {
     instance = axios.create({
       baseURL: baseUrl,
@@ -79,11 +68,7 @@ module.exports = ({ baseUrl, rateLimiter }) => {
     })
 
     instance.interceptors.response.use((res) => {
-      if (res.headers.hasOwnProperty('set-cookie')) {
-        for (const cookie of res.headers['set-cookie']) {
-          setCookie(cookie.split('=')[0], cookie.split('=')[1].split(';')[0])
-        }
-      }
+      debug(res.headers)
       return res
     }, (error) => {
       if (error.response) {
@@ -107,8 +92,6 @@ module.exports = ({ baseUrl, rateLimiter }) => {
     _get,
     _post,
     _put,
-    _delete,
-    setCookie,
-    deleteCookie
+    _delete
   }
 }
